Changes in LED_Driver.c
Changes in delay.c (unsinged long int)
add for loop in led_driver.c

+++ source/delay.c
@@ -1,8 +1,8 @@
 /* file for creating delay */
 #include "../include/delay.h"
 
-void delay_us(int us)	   	// Delay function
+void delay_us(unsigned long int us)	   	// Delay function
 	{
-	  int i;
+	  unsigned long int i;
 	  for (i=0; i<us; i++);
 	}
